+++ crates/ryo3-core/src/lib.rs
@@ -1,2 +1,5 @@
+mod errors;
 pub mod pystring;
 pub mod types;
+
+pub use errors::FeatureNotEnabledError;