+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## v0.0.34 [unreleased]
+
+- `ryo3-std`
+  - `fs`:
+    - `read_stream` function that returns an iterator of `ry.Bytes` objects from a `PathLike` object
+    - Several more fs functions added
+- `ryo3-tokio`
+  - Several more tokio fs functions added
+- internal
+  - reorganized type annotations to be not a HUGE file...
+
+___
+
 ## v0.0.33 [2025-02-26]
 
 - update to pyo3 v0.23.5